lstm

+++ seek/text_dataset.py
@@ -5,8 +5,8 @@ import torch
 
 class TextDataset(Dataset):
     def __init__(self, data, label, trans=None):
-        self.dt = torch.asarray(data, dtype=torch.float32)
-        self.lb = torch.asarray(label, dtype=torch.long)
+        self.dt = torch.asarray(data, dtype=torch.int32)
+        self.lb = torch.asarray(label, dtype=torch.float32)
         self.trans = trans
 
     def __len__(self):
